feat: restructure, add types in local-tools
