+++ composio/local_tools/greptile/__init__.py
@@ -1,2 +1,2 @@
-from .codequery import CodeQuery
+from .actions.codequery import CodeQuery
 from .tool import Greptile
