+++ composio/local_tools/mathematical/tool.py
@@ -1,13 +1,13 @@
-from ..tool import Tool
-from .calculator import Calculator
-
+from ..tool import Tool, Action
+from .actions.calculator import Calculator
+import typing as t
 
 class Mathematical(Tool):
     """
     Mathematical Tools for LLM
     """
 
-    def actions(self) -> list:
+    def actions(self) -> list[t.Type[Action]]:
         return [Calculator]
 
     def triggers(self) -> list:
