+++ composio/local_tools/ragtool/actions/rag_query.py
@@ -1,27 +1,11 @@
 from pydantic import BaseModel, Field
-from typing import Any, Type, List
-from ..action import Action
-from composio.local_tools.tool import Tool
-from ..tool import Tool
-
+from ...tool import Action
 
 class RagToolQueryRequest(BaseModel):
     query: str = Field(..., description="The query to search in the knowledge base")
 
 class RagToolQueryResponse(BaseModel):
     response: str = Field(..., description="The response to the query from the knowledge base")
-
-class RagToolAddRequest(BaseModel):
-    content: str = Field(..., description="Content to add to the knowledge base")
-
-class RagToolAddResponse(BaseModel):
-    status: str = Field(..., description="Status of the addition to the knowledge base")
-
-class RagToolResetRequest(BaseModel):
-    confirmation: bool = Field(..., description="Confirmation to reset the state")
-
-class RagToolResetResponse(BaseModel):
-    status: str = Field(..., description="Status of the reset action")
 
 class RagToolQuery(Action):
     """
@@ -60,34 +44,3 @@ class RagToolQuery(Action):
                 return f"Error querying knowledge base: {e}"
         else:
             return "App initialization failed, cannot perform query."
-
-class AddContentToRagTool(Action):
-    """
-    Tool for adding content to the knowledge base
-    """
-
-    _display_name = "Add Content to Rag Tool"
-    _request_schema = RagToolAddRequest
-    _response_schema = RagToolAddResponse
-    _tags = ["Knowledge Base"]
-    _tool_name = "ragtoolactions"
-
-    def execute(self, request: RagToolAddRequest, authorisation_data: dict = {}):
-        """Add content to the knowledge base"""
-        try:
-            from embedchain import App
-        except ImportError as e:
-            raise ImportError(f"Failed to import App from embedchain: {e}")
-        try:
-            embedchain_app = App()
-            content = request.content
-            embedchain_app.add(content)
-            return "Content added successfully"
-        except Exception as e:
-            raise Exception(f"Error adding content: {e}")
-
-class RagToolActions(Tool):
-    """Rag Tool Actions"""
-
-    def actions(self):
-        return [RagToolQuery, AddContentToRagTool]
